+++ glean-core/ios/Glean/Glean.swift
@@ -355,12 +355,14 @@ public class Glean {
 
     /// Test-only method to destroy the owned glean-core handle.
     func testDestroyGleanHandle(_ clearStores: Bool = false) {
+        // If it was initialized this also clears the directory
+        let dataPath = getGleanDirectory().relativePath
+        gleanTestDestroyGlean(clearStores, dataPath)
+
         if !isInitialized() {
-            // We don't need to destroy Glean: it wasn't initialized.
+            // We don't need to destroy anything else: it wasn't initialized.
             return
         }
-
-        gleanTestDestroyGlean(clearStores)
 
         // Reset all state
         gleanSetTestMode(false)
@@ -392,15 +394,6 @@ public class Glean {
                            uploadEnabled: Bool = true) {
         // Init Glean.
         testDestroyGleanHandle(clearStores)
-
-        // If Glean was not initialized we can't ask it to delete its data.
-        // But we can go as far as simply removing all files.
-        if clearStores {
-            let dataPath = getGleanDirectory()
-            // Don't care if we can't remove the directory.
-            // It might just not exist.
-            _ = try? FileManager.default.removeItem(at: dataPath)
-        }
 
         // Enable test mode.
         enableTestingMode()